Fix grapheme cluster examples
There was a problem getting an emoji through both the syntax
highlighting and the html encoding.

If an emoji were entered directly into the content, it was passed through the highlighter and not escaped. So real bug was in the
highlighter.

If the unicode value was entered into the content, then the leading
ampersand was being escaped and the code would just print out.

Fixed by adding a ReplaceRE to swap out the escaped unicode for
the non escaped value for one specific unicode character used in the
example blocks.